Move the data sourcing actor to the `microgrid` module
