+++ src/frequenz/sdk/microgrid/_data_sourcing/_component_metric_request.py
