+++ src/frequenz/sdk/microgrid/_data_sourcing/data_sourcing.py
@@ -6,7 +6,7 @@
 from frequenz.channels import Receiver
 
 from ..._internal._channels import ChannelRegistry
-from .._actor import Actor
+from ...actor import Actor
 from ._component_metric_request import ComponentMetricRequest
 from .microgrid_api_source import MicrogridApiSource
