+++ src/frequenz/sdk/microgrid/_resampling.py
@@ -13,10 +13,10 @@ from frequenz.channels import Receiver, Sender
 from .._internal._asyncio import cancel_and_await
 from .._internal._channels import ChannelRegistry
 from ..actor import Actor
-from ..actor._data_sourcing import ComponentMetricRequest
 from ..timeseries import Sample
 from ..timeseries._quantities import Quantity
 from ..timeseries._resampling import Resampler, ResamplerConfig, ResamplingError
+from ._data_sourcing import ComponentMetricRequest
 
 _logger = logging.getLogger(__name__)
