+++ src/frequenz/sdk/timeseries/consumer.py
@@ -8,7 +8,7 @@ import uuid
 from frequenz.channels import Sender
 
 from .._internal._channels import ChannelRegistry
-from ..actor import ComponentMetricRequest
+from ..microgrid._data_sourcing import ComponentMetricRequest
 from ._quantities import Power
 from .formula_engine import FormulaEngine
 from .formula_engine._formula_engine_pool import FormulaEnginePool
