+++ src/frequenz/sdk/timeseries/formula_engine/_formula_engine_pool.py
@@ -11,6 +11,7 @@ from frequenz.channels import Sender
 from frequenz.client.microgrid import ComponentMetricId
 
 from ..._internal._channels import ChannelRegistry
+from ...microgrid._data_sourcing import ComponentMetricRequest
 from .._quantities import Current, Power, Quantity
 from ._formula_generators._formula_generator import (
     FormulaGenerator,
@@ -20,7 +21,6 @@ from ._resampled_formula_builder import ResampledFormulaBuilder
 
 if TYPE_CHECKING:
     # Break circular import
-    from ...actor import ComponentMetricRequest
     from ..formula_engine import FormulaEngine, FormulaEngine3Phase
