+++ src/frequenz/sdk/timeseries/grid.py
@@ -11,13 +11,13 @@ from __future__ import annotations
 import logging
 import uuid
 from dataclasses import dataclass
-from typing import TYPE_CHECKING
 
 from frequenz.channels import Sender
 from frequenz.client.microgrid._component import ComponentCategory
 
 from .._internal._channels import ChannelRegistry
 from ..microgrid import connection_manager
+from ..microgrid._data_sourcing import ComponentMetricRequest
 from ._fuse import Fuse
 from ._quantities import Current, Power
 from .formula_engine import FormulaEngine, FormulaEngine3Phase
@@ -27,10 +27,6 @@ from .formula_engine._formula_generators import (
     GridPower3PhaseFormula,
     GridPowerFormula,
 )
-
-if TYPE_CHECKING:
-    # Break circular import
-    from ..actor import ComponentMetricRequest
 
 _logger = logging.getLogger(__name__)
