+++ src/frequenz/sdk/timeseries/pv_pool/_pv_pool_reference_store.py
@@ -12,8 +12,8 @@ from frequenz.channels import Broadcast, Receiver, Sender
 from frequenz.client.microgrid import ComponentCategory, InverterType
 
 from ..._internal._channels import ChannelRegistry, ReceiverFetcher
-from ...actor import ComponentMetricRequest
 from ...microgrid import connection_manager
+from ...microgrid._data_sourcing import ComponentMetricRequest
 from ...microgrid._power_distributing import ComponentPoolStatus, Result
 from ...microgrid._power_managing._base_classes import Proposal, ReportRequest
 from .._base_types import SystemBounds
