+++ tests/timeseries/mock_resampler.py
@@ -13,9 +13,9 @@ from frequenz.client.microgrid import ComponentMetricId
 from pytest_mock import MockerFixture
 
 from frequenz.sdk._internal._asyncio import cancel_and_await
-from frequenz.sdk.actor import ComponentMetricRequest, ResamplerConfig
 from frequenz.sdk.microgrid._data_pipeline import _DataPipeline
-from frequenz.sdk.timeseries import Sample
+from frequenz.sdk.microgrid._data_sourcing import ComponentMetricRequest
+from frequenz.sdk.timeseries import ResamplerConfig, Sample
 from frequenz.sdk.timeseries._quantities import Quantity
 from frequenz.sdk.timeseries.formula_engine._formula_generators._formula_generator import (
     NON_EXISTING_COMPONENT_ID,
